bookmarks

+++ lib/components/add_post/add_post_model.dart
@@ -14,7 +14,7 @@ import 'package:path_provider/path_provider.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:stop_watch_timer/stop_watch_timer.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:whisper/constants/doubles.dart';
+import 'package:whisper/constants/nums.dart';
 import 'package:whisper/constants/strings.dart';
 // constants
 import 'package:whisper/constants/lists.dart';
@@ -238,7 +238,7 @@ class AddPostModel extends ChangeNotifier {
       playCount: 0,
       postId: postId, 
       positiveScore: 0,
-      score: defaultScore,
+      score: defaultScore.toDouble(),
       storageImageName: storageImageName, 
       storagePostName: storagePostName, 
       tagUids: [],
